+++ frontend/src/views/Settings.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import { computed, ref } from 'vue';
+import { computed } from 'vue';
 import { useUserStore } from '@/stores/user';
 import type { User } from '@/types';
 import api from '@/api';
@@ -8,26 +8,15 @@ import SettingsForm from '@/components/SettingsForm.vue';
 const store = useUserStore()
 const currentUser = computed<User | null>(() => store.getCurrentUser)
 
-const currentLanguage = localStorage.getItem('language') ?? 'en'
-
-const language = ref(currentLanguage)
-
 const updateBeta = async (beta: boolean) => {
   await api.put<User>('settings', { beta })
   store.setCurrentUser()
 }
 
-const updateLanguage = () => {
-  localStorage.setItem('language', language.value)
-}
 </script>
 
 <template>
   <div>
     <SettingsForm v-if="currentUser" :user="currentUser" :updateBeta="updateBeta" />
-    <select v-model="language" @change="updateLanguage">
-      <option value="en">English</option>
-      <option value="it">Italian</option>
-    </select>
   </div>
 </template>
